requirements.txt
sqlalchemy
ipython
psycopg2
pytest
python-dotenv
